:truck: Move app component to app directory
