+++ src/app/App.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import './App.css';
-import Header from "./components/Header";
-import FileGroup from "./components/FileGroup";
-import {data} from "./Metadata";
+import Header from "../components/Header";
+import FileGroup from "../components/FileGroup";
+import {data} from "../Metadata";
 import {createMuiTheme, MuiThemeProvider} from '@material-ui/core/styles';
-import Jumbotron from "./components/Jumbotron";
-import Acknowledgement from "./components/Acknowledgement";
+import Jumbotron from "../components/Jumbotron";
+import Acknowledgement from "../components/Acknowledgement";
 
 function App() {
     const appTheme = createMuiTheme({
